move tour models out of auth to their own file
